CR 6655847: After starting a test, AutoTester in MVM must wait until *all* isolates are finished
Part 2.
These changes are needed to test CHAPI.
All events that are in the queue at the moment when the test isolate has exited are handled (because when testing
CHAPI there may be requests to start new isolates). To do this, AutoTester sends AUTO_TEST event to himself;
when this event arrives, all events that were placed in the queue before it are guaranteed to be processed. 

svn merge -r9744:10065 ../branches/midp-cr-6655847-part2

Reviewed by Alexey Kornev.